refactor: modularize project structure
