+++ src/modules/economy/commands/balance.js
