+++ src/modules/economy/commands/daily.js
