+++ src/modules/economy/commands/earn.js
