+++ src/modules/economy/commands/weekly.js
