+++ src/modules/games/commands/dice.js
