+++ src/modules/games/commands/rps.js
