+++ src/modules/moderation/commands/mute.js
